+++ test/test_bit_read_messages.py
@@ -45,7 +45,7 @@ class ModbusBitMessageTests(unittest.TestCase):
         msg    = "ReadBitRequest(1,1)"
         self.assertEqual(msg, str(handle))
         handle = ReadBitsResponseBase([1,1])
-        msg    = "ReadBitResponse(2)"
+        msg    = "ReadBitsResponseBase(2)"
         self.assertEqual(msg, str(handle))
 
     def testBitReadBaseRequestEncoding(self):