+++ test/test_server_asyncio.py
@@ -72,6 +72,7 @@ class AsyncioServerTest(asynctest.TestCase):
     # Test ModbusConnectedRequestHandler
     #-----------------------------------------------------------------------#
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testStartTcpServer(self):
         ''' Test that the modbus tcp asyncio server starts correctly '''
         identity = ModbusDeviceIdentification(info={0x00: 'VendorName'})
@@ -99,6 +100,7 @@ class AsyncioServerTest(asynctest.TestCase):
             serve.assert_awaited()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerServeForeverTwice(self):
         ''' Call on serve_forever() twice should result in a runtime error '''
         server = yield from StartTcpServer(context=self.context,address=("127.0.0.1", 0), loop=self.loop)
@@ -112,6 +114,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerReceiveData(self):
         ''' Test data sent on socket is received by internals - doesn't not process data '''
         data = b'\x01\x00\x00\x00\x00\x06\x01\x03\x00\x00\x00\x19'
@@ -146,6 +149,7 @@ class AsyncioServerTest(asynctest.TestCase):
             server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerRoundtrip(self):
         ''' Test sending and receiving data on tcp socket '''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x03\x00\x00\x00\x01" # unit 1, read register
@@ -186,6 +190,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerConnectionLost(self):
         ''' Test tcp stream interruption '''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x01\x00\x00\x00\x01"
@@ -224,6 +229,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerCloseActiveConnection(self):
         ''' Test server_close() while there are active TCP connections '''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x01\x00\x00\x00\x01"
@@ -259,6 +265,7 @@ class AsyncioServerTest(asynctest.TestCase):
         self.assertTrue( len(server.active_connections) == 0 )
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerNoSlave(self):
         ''' Test unknown slave unit exception '''
         context = ModbusServerContext(slaves={0x01: self.store, 0x02: self.store  }, single=False)
@@ -295,6 +302,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerModbusError(self):
         ''' Test sending garbage data on a TCP socket should drop the connection '''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x03\x00\x00\x00\x01"  # get slave 5 function 3 (holding register)
@@ -334,6 +342,7 @@ class AsyncioServerTest(asynctest.TestCase):
             server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerInternalException(self):
         ''' Test sending garbage data on a TCP socket should drop the connection '''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x03\x00\x00\x00\x01"  # get slave 5 function 3 (holding register)
@@ -378,6 +387,7 @@ class AsyncioServerTest(asynctest.TestCase):
     # Test ModbusTlsProtocol
     #-----------------------------------------------------------------------#
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testStartTlsServer(self):
         ''' Test that the modbus tls asyncio server starts correctly '''
         with patch.object(ssl.SSLContext, 'load_cert_chain') as mock_method:
@@ -409,6 +419,7 @@ class AsyncioServerTest(asynctest.TestCase):
                 serve.assert_awaited()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTlsServerServeForeverTwice(self):
         ''' Call on serve_forever() twice should result in a runtime error '''
         with patch.object(ssl.SSLContext, 'load_cert_chain') as mock_method:
@@ -428,6 +439,7 @@ class AsyncioServerTest(asynctest.TestCase):
     #-----------------------------------------------------------------------#
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testStartUdpServer(self):
         ''' Test that the modbus udp asyncio server starts correctly '''
         identity = ModbusDeviceIdentification(info={0x00: 'VendorName'})
@@ -454,6 +466,7 @@ class AsyncioServerTest(asynctest.TestCase):
             serve.assert_awaited()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerServeForeverClose(self):
         ''' Test StartUdpServer serve_forever() method '''
         server = yield from StartUdpServer(context=self.context,address=("127.0.0.1", 0), loop=self.loop)
@@ -470,6 +483,7 @@ class AsyncioServerTest(asynctest.TestCase):
         self.assertTrue(server.protocol.is_closing())
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerServeForeverTwice(self):
         ''' Call on serve_forever() twice should result in a runtime error '''
         identity = ModbusDeviceIdentification(info={0x00: 'VendorName'})
@@ -485,6 +499,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerReceiveData(self):
         ''' Test that the sending data on datagram socket gets data pushed to framer '''
         server = yield from StartUdpServer(context=self.context,address=("127.0.0.1", 0),loop=self.loop)
@@ -506,6 +521,7 @@ class AsyncioServerTest(asynctest.TestCase):
             server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerSendData(self):
         ''' Test that the modbus udp asyncio server correctly sends data outbound '''
         identity = ModbusDeviceIdentification(info={0x00: 'VendorName'})
@@ -548,6 +564,7 @@ class AsyncioServerTest(asynctest.TestCase):
         yield from asyncio.sleep(0.1)
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerRoundtrip(self):
         ''' Test sending and receiving data on udp socket'''
         data = b"\x01\x00\x00\x00\x00\x06\x01\x03\x00\x00\x00\x01" # unit 1, read register
@@ -586,6 +603,7 @@ class AsyncioServerTest(asynctest.TestCase):
         server.server_close()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testUdpServerException(self):
         ''' Test sending garbage data on a TCP socket should drop the connection '''
         garbage = b'\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF'
@@ -624,16 +642,19 @@ class AsyncioServerTest(asynctest.TestCase):
     # -----------------------------------------------------------------------#
     # Test ModbusServerFactory
     # -----------------------------------------------------------------------#
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testModbusServerFactory(self):
         ''' Test the base class for all the clients '''
         with self.assertWarns(DeprecationWarning):
             factory = ModbusServerFactory(store=None)
 
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testStopServer(self):
         with self.assertWarns(DeprecationWarning):
             StopServer()
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerException(self):
         ''' Sending garbage data on a TCP socket should drop the connection '''
         garbage = b'\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF'
@@ -674,6 +695,7 @@ class AsyncioServerTest(asynctest.TestCase):
 
 
     @asyncio.coroutine
+    @pytest.mark.skipif(not IS_PYTHON3, reason="requires python3.4 or above")
     def testTcpServerException(self):
         ''' Sending garbage data on a TCP socket should drop the connection '''
         garbage = b'\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF'